fixed conversion

+++ libft/ft_anylen.c
@@ -6,7 +6,7 @@
 /*   By: altikka <altikka@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/04 12:42:13 by altikka           #+#    #+#             */
-/*   Updated: 2022/05/04 13:49:29 by altikka          ###   ########.fr       */
+/*   Updated: 2022/05/08 22:31:46 by altikka          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ size_t	ft_anylen(uintmax_t n, int base, int sign)
 		len++;
 	while (n != 0)
 	{
-		n /= base;
+		n /= (unsigned long ) base;
 		len++;
 	}
 	return (len);
