+++ libft/ft_vecappend.c
@@ -6,7 +6,7 @@
 /*   By: altikka <altikka@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/27 13:34:14 by altikka           #+#    #+#             */
-/*   Updated: 2022/04/27 14:33:24 by altikka          ###   ########.fr       */
+/*   Updated: 2022/05/08 22:29:54 by altikka          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ int	ft_vecappend(t_vec *dest, t_vec *src)
 {
 	size_t	new_size;
 	int		ret;
-	int		i;
+	size_t	i;
 
 	if (!dest || !src)
 		return (-1);
